requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
#
#
python=3.8.12=h12debd9_0
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
cudatoolkit=11.3.1=h2bc3f7f_2
torchvision=0.12.0=py38_cu113
numpy=1.21.2=py38h20f2e39_0
pillow=9.0.1=py38h22f2fdc_0
pip=21.2.4=py38h06a4308_0
timm=0.4.12=pyhd8ed1ab_0
tqdm=4.64.0=py38h06a4308_0
scikit-learn=1.0.2=py38h51133e4_1